requirements.txt
pyserial
pandas
scikit-learn
matplotlib